Add Author user control
