+++ Library/Controls/Mode.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace Library.Controls
+{
+    public enum Mode
+    {
+        Add,
+        Edit
+    }
+}
