+++ Library/Web.config
@@ -11,6 +11,11 @@
   <system.web>
     <compilation debug="true" targetFramework="4.8"/>
     <httpRuntime targetFramework="4.8"/>
+    <pages>
+      <controls>
+        <add tagPrefix="uc" src="~/Controls/Author.ascx" tagName="Author" />
+      </controls>
+    </pages>
   </system.web>
   <system.codedom>
     <compilers>
